# Milestone 4/backend/swagger.yaml
openapi: 3.1.0
info:
  title: EduAssist API
  description: AI-Powered Teaching Assistant Platform - Knowledge Assistant Feature
  version: 1.0.0
paths:
  /knowledge-assistant/knowledge/documents:
    post:
      tags:
        - Knowledge Assistant
        - Knowledge Assistant
      summary: Upload Document
      description: Upload and process a document for RAG knowledge base
      operationId: upload_document_knowledge_assistant_knowledge_documents_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: title
          in: query
          required: false
          schema:
            type: string
            title: Title
        - name: course_id
          in: query
          required: false
          schema:
            type: integer
            default: 1
            title: Course Id
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_document_knowledge_assistant_knowledge_documents_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Knowledge Assistant
        - Knowledge Assistant
      summary: List Documents
      description: List all documents
      operationId: list_documents_knowledge_assistant_knowledge_documents_get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: course_id
          in: query
          required: false
          schema:
            type: integer
            title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentResponse'
                title: Response List Documents Knowledge Assistant Knowledge Documents Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /knowledge-assistant/knowledge/conversations:
    post:
      tags:
        - Knowledge Assistant
        - Knowledge Assistant
      summary: Create Conversation
      description: Create a new conversation
      operationId: create_conversation_knowledge_assistant_knowledge_conversations_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OAuth2PasswordBearer: []
  /knowledge-assistant/knowledge/conversations/{conversation_id}:
    get:
      tags:
        - Knowledge Assistant
        - Knowledge Assistant
      summary: Get Conversation
      description: Get conversation with messages
      operationId: get_conversation_knowledge_assistant_knowledge_conversations__conversation_id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: conversation_id
          in: path
          required: true
          schema:
            type: integer
            title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /knowledge-assistant/knowledge/chat:
    post:
      tags:
        - Knowledge Assistant
        - Knowledge Assistant
      summary: Chat Query
      description: |-
        Process a chat query using RAG

        Steps:
        1. Query knowledge base for relevant context
        2. Generate response with Gemini
        3. Save message to conversation
        4. Return response with confidence score
      operationId: chat_query_knowledge_assistant_knowledge_chat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatQueryRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatQueryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OAuth2PasswordBearer: []
...
  /feedback/:
    post:
      tags:
        - Feedback
      summary: Submit Feedback
      operationId: submit_feedback_feedback__post
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            default: 1
            title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedbackCreate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedbackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
        - Feedback
      summary: Get Feedback
      operationId: get_feedback_feedback__get
      parameters:
        - name: target_id
          in: query
          required: true
          schema:
            type: integer
            title: Target Id
        - name: target_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/FeedbackTargetType'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedbackResponse'
                title: Response Get Feedback Feedback  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/register:
    post:
      tags:
        - Auth
      summary: Register
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Body_login_auth_login_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/password-reset-request:
    post:
      tags:
        - Auth
      summary: Password Reset Request
      operationId: password_reset_request_auth_password_reset_request_post
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            title: Email
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/reset-password:
    post:
      tags:
        - Auth
      summary: Password Reset
      operationId: password_reset_auth_reset_password_post
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            title: Token
        - name: new_password
          in: query
          required: true
          schema:
            type: string
            title: New Password
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /register:
    post:
      summary: Register
      description: Register a new user
      operationId: register_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_login_auth_login_post:
      properties:
        grant_type:
          anyOf:
            - type: string
              pattern: password
            - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Secret
      type: object
      required:
        - username
        - password
      title: Body_login_auth_login_post
...
    Body_upload_document_knowledge_assistant_knowledge_documents_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
        - file
      title: Body_upload_document_knowledge_assistant_knowledge_documents_post
    ChatQueryRequest:
      properties:
        query:
          type: string
          maxLength: 5000
          minLength: 1
          title: Query
        course_id:
          type: integer
          title: Course Id
        conversation_id:
          anyOf:
            - type: integer
            - type: 'null'
          title: Conversation Id
      type: object
      required:
        - query
        - course_id
      title: ChatQueryRequest
    ChatQueryResponse:
      properties:
        message:
          $ref: '#/components/schemas/MessageResponse'
        should_escalate:
          type: boolean
          title: Should Escalate
        confidence_score:
          type: number
          title: Confidence Score
      type: object
      required:
        - message
        - should_escalate
        - confidence_score
      title: ChatQueryResponse
    ConversationCreate:
      properties:
        course_id:
          type: integer
          title: Course Id
      type: object
      required:
        - course_id
      title: ConversationCreate
    ConversationResponse:
      properties:
        id:
          type: integer
          title: Id
        student_id:
          type: integer
          title: Student Id
        course_id:
          type: integer
          title: Course Id
        status:
          $ref: '#/components/schemas/ConversationStatus'
        started_at:
          type: string
          format: date-time
          title: Started At
        last_message_at:
          anyOf:
            - type: string
              format: date-time
            - type: 'null'
          title: Last Message At
        messages:
          items:
            $ref: '#/components/schemas/MessageResponse'
          type: array
          title: Messages
          default: []
      type: object
      required:
        - id
        - student_id
        - course_id
        - status
        - started_at
      title: ConversationResponse
    ConversationStatus:
      type: string
      enum:
        - ACTIVE
        - RESOLVED
        - ESCALATED
      title: ConversationStatus
...
    DocumentResponse:
      properties:
        id:
          type: integer
          title: Id
        title:
          type: string
          title: Title
        file_path:
          type: string
          title: File Path
        file_type:
          type: string
          title: File Type
        file_size:
          type: integer
          title: File Size
        processed:
          type: boolean
          title: Processed
        course_id:
          type: integer
          title: Course Id
        uploaded_by:
          type: integer
          title: Uploaded By
        uploaded_at:
          type: string
          format: date-time
          title: Uploaded At
      type: object
      required:
        - id
        - title
        - file_path
        - file_type
        - file_size
        - processed
        - course_id
        - uploaded_by
        - uploaded_at
      title: DocumentResponse
    FeedbackCreate:
      properties:
        target_id:
          type: integer
          title: Target Id
        target_type:
          $ref: '#/components/schemas/FeedbackTargetType'
        rating:
          type: number
          maximum: 5
          minimum: 1
          title: Rating
        aspect_ratings:
          anyOf:
            - additionalProperties:
                type: number
              type: object
            - type: 'null'
          title: Aspect Ratings
        comment:
          anyOf:
            - type: string
            - type: 'null'
          title: Comment
      type: object
      required:
        - target_id
        - target_type
        - rating
      title: FeedbackCreate
    FeedbackResponse:
      properties:
        target_id:
          type: integer
          title: Target Id
        target_type:
          $ref: '#/components/schemas/FeedbackTargetType'
        rating:
          type: number
          maximum: 5
          minimum: 1
          title: Rating
        aspect_ratings:
          anyOf:
            - additionalProperties:
                type: number
              type: object
            - type: 'null'
          title: Aspect Ratings
        comment:
          anyOf:
            - type: string
            - type: 'null'
          title: Comment
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        created_at:
          type: string
          title: Created At
      type: object
      required:
        - target_id
        - target_type
        - rating
        - id
        - user_id
        - created_at
      title: FeedbackResponse
    FeedbackTargetType:
      type: string
      enum:
        - assessment
        - study_guide
        - slide_deck
      title: FeedbackTargetType
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    MessageResponse:
      properties:
        id:
          type: integer
          title: Id
        conversation_id:
          type: integer
          title: Conversation Id
        sender_type:
          $ref: '#/components/schemas/SenderType'
        content:
          type: string
          title: Content
        confidence_score:
          anyOf:
            - type: number
            - type: 'null'
          title: Confidence Score
        sources:
          anyOf:
            - type: object
            - type: 'null'
          title: Sources
        timestamp:
          type: string
          format: date-time
          title: Timestamp
      type: object
      required:
        - id
        - conversation_id
        - sender_type
        - content
        - timestamp
      title: MessageResponse
...
    SenderType:
      type: string
      enum:
        - STUDENT
        - AI
        - TA
      title: SenderType
...
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        full_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Full Name
        password:
          type: string
          title: Password
      type: object
      required:
        - email
        - username
        - password
      title: UserCreate
    UserResponse:
      properties:
        email:
          type: string
          format: email
          title: Email
        username:
          type: string
          title: Username
        full_name:
          anyOf:
            - type: string
            - type: 'null'
          title: Full Name
        id:
          type: integer
          title: Id
        is_active:
          type: boolean
          title: Is Active
        is_student:
          type: boolean
          title: Is Student
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
        - email
        - username
        - id
        - is_active
        - is_student
        - created_at
      title: UserResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token